Main program
am LAZY. I need to create a program for calculating current bill for a two-storey house, which is given for rent, in which one of each floor is used by different families. So obviously the problem comes when sharing the current bill. It cant be shared equally as the number of appliances are different for both families. The upper floor has a seperate current meter which measures the number of units used. But the official connection with KSEB is for the house as whole. This means the bill shows the total units used by the whole house only. To divide the payable bill amount to both the floors simple calculations must be made. But that has to be automated because, I AM LAZY.